cli: Print newly tracked files
Tracking file automatically is a neat thing, until you did not want to
track the auto-tracked file. This could happen at any jj command line.
You can then, easily end-up having unwanted files in your commits,
without realizing you have them.

This commits makes so a message is printed to indicate when jj started
auto-tracking some files. With the intent to have a quick response
from the user when one of the newly tracked files is not one they want
to have tracked.

The message here is limited to one single line, we do not want to spam
the output. It replicates the same type of message that `jj file
untrack` does when attempting to untrack a non-ignored file, with
auto-tracking enabled.

So the message does not list all tracked files, but hopefully it will
raise enough questions so the user checks the content of its commit.