cli: Print newly tracked files
XXX New idea:
- Have the snapshot stats provide the list of newly tracked paths,
  untracked paths and ignored path as already collapsed.
- Allows to collapse + sort properly. Then the display is much more
  simple.

Tracking file automatically is a neat thing, until you did not want to
track the file that just got auto-tracked. This could happen on many
jj commands, so you could, easily end-up having unwanted files in your
commits, without realizing you have them.

We are now printing a message to indicate when jj starts auto-tracking
files. With the intent to have a quick response from the user when one
of the newly tracked files is not one they want to have tracked.

The message to not take more one single line of your terminal, to not
spam. It will list the path of as much newly tracked files as possible,
and giving the actual number of files it could not print.

Hopefully it should trigger the user into running another command
(such as `jj status`) to check what is the state of the commit. And
take appropriate action to ignore and untrack the unwanted files.